added sidekiq and send survey job
