+++ .gitignore
@@ -36,7 +36,8 @@ yarn-debug.log*
 .yarn-integrity
 # Ignore .env file containing credentials.
 .env*
+ngrok.yml
 # Ignore Mac and Linux file system files
 *.swp
 .DS_Store
-.env*
+
