+++ Gemfile
@@ -22,6 +22,8 @@ gem 'redis', '~> 4.0'
 gem "chartkick"
 gem 'groupdate'
 gem 'hashie'
+gem 'sidekiq'
+gem 'sidekiq-failures', '~> 1.0'
 # Use Active Model has_secure_password
 # gem 'bcrypt', '~> 3.1.7'
