+++ config/application.rb
@@ -6,6 +6,8 @@ require 'rails/all'
 # you've limited to :test, :development, or :production.
 Bundler.require(*Rails.groups)
 
+
+
 module Kyoro
   class Application < Rails::Application
     config.generators do |generate|
@@ -15,7 +17,7 @@ module Kyoro
     end
     # Initialize configuration defaults for originally generated Rails version.
     config.load_defaults 6.0
-
+    config.active_job.queue_adapter = :sidekiq
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration can go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded after loading
