+++ config/routes.rb
@@ -4,6 +4,10 @@ Rails.application.routes.draw do
   root to: 'pages#home'
   # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
 
+  require "sidekiq/web"
+  authenticate :user, ->(user) { user.admin? } do
+    mount Sidekiq::Web => '/sidekiq'
+  end
 
   resources :users, only: [:index, :show, :edit, :update]
   resources :surveys, only: [:index, :show, :new, :create ]
