+++ ngrok.yml
@@ -1,1 +1,1 @@
-
+authtoken: 1tW9mQxXTl73RLfN2zpv5Yn1j3x_2Jgntr6tj5FBjDwb4ZGw4
