+++ test/jobs/send_survey_job_test.rb
@@ -1,0 +1,7 @@
+require 'test_helper'
+
+class SendSurveyJobTest < ActiveJob::TestCase
+  # test "the truth" do
+  #   assert true
+  # end
+end
